making getOpenOrders a utility function
